requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.9.1
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clickhouse-connect==0.8.15
clickhouse-driver==0.2.9
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
ijson==3.3.0
Jinja2==3.1.6
lxml==5.3.1
lz4==4.4.3
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.2.0
numpy==2.0.2
pandas==2.2.3
pendulum==3.0.0
propcache==0.3.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
